+++ src/Components/App.module.css
@@ -1,0 +1,11 @@
+body{
+    background-color: aqua;
+}
+
+.heading{
+    text-align: center;
+}
+
+.img{
+    text-align: center;
+}
