+++ src/index.js
@@ -1,8 +1,5 @@
 import React from "react";
 import reactDOM from "react-dom";
-
-const App = () => {
-    return <h1>Hello From React</h1>
-};
+import App from './Components/App.js'
 
 reactDOM.render(<App/>,document.getElementById('root'))
