+++ webpack.config.js
@@ -1,28 +1,7 @@
-const path = require("path");
-const HtmlWebpackPlugin = require("html-webpack-plugin");
-module.exports = {
-  entry: "./src/index.js",
-  output: {
-    filename: "bundle.js",
-    path: path.join(__dirname, "dist"),
-  },
-  module: {
-    rules: [
-      {
-        exclude: /node_modules/,
-        test: /\.js$|jsx/,
-        use: {
-          loader: "babel-loader",
-          options: {
-            presets: ["@babel/preset-env", "@babel/preset-react"],
-          },
-        },
-      },
-    ],
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-    template: path.join(__dirname, "src", "index.html"),
-    }),
-    ],
+const { merge } = require("webpack-merge");
+const commonWebpackConfig = require("./webpack.common.js");
+ 
+module.exports = ({ ENV }) => {
+ const envWebpackConfig = require(`./webpack.${ENV}.js`);
+ return merge(commonWebpackConfig, envWebpackConfig);
 };
